projects/Poetry Analyzer with chainlit/pyproject.toml
[project]
name = "poetry-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.0",
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]
